Only logged in users can now post or comment and users can only delete their own posts/comments

+++ server/repositories/commentRepository.js
@@ -2,10 +2,10 @@ import postgres from "postgres";
 
 const sql = postgres();
 //ppId = Parent Post Id
-const create = async (communityId, ppId, comment) => {          
+const create = async (userId, communityId, ppId, comment) => {          
     const result = await sql`
-        INSERT INTO posts (community_id, content, parent_post_id)
-        VALUES (${communityId}, ${comment.content}, ${ppId})                                    
+        INSERT INTO posts (community_id, content, parent_post_id, created_by)
+        VALUES (${communityId}, ${comment.content}, ${ppId}, ${userId})                                    
         RETURNING *;`;
     return result[0];                         
 };
@@ -14,10 +14,13 @@ const findAll = async (communityId, ppId) => {
         SELECT * FROM posts
         WHERE community_id = ${communityId} AND parent_post_id = ${ppId}`;
 };
-const deleteComment = async (communityId, ppId, commentId) => {
+const deleteComment = async (userId, communityId, ppId, commentId) => {
     const result = await sql`
         DELETE FROM posts
-        WHERE id = ${commentId} AND community_id = ${communityId} AND parent_post_id = ${ppId}
+        WHERE id = ${commentId} 
+            AND community_id = ${communityId} 
+            AND parent_post_id = ${ppId}
+            AND created_by = ${userId}
         RETURNING *;`;
     if (result.length === 0) {
         return undefined;
